Fixed wrong return type [skip ci]

+++ src/pocketmine/entity/behavior/navigator/algorithms/DistanceAlgorithm.php
@@ -26,5 +26,5 @@ use pocketmine\utils\navigator\Tile;
 
 interface DistanceAlgorithm{
 	
-	public function calculate(Tile $from, Tile $to) : int;
+	public function calculate(Tile $from, Tile $to) : float;
 }
