+++ src/pocketmine/entity/behavior/navigator/algorithms/ManhattanHeuristicAlgorithm.php
@@ -26,7 +26,7 @@ use pocketmine\utils\navigator\Tile;
 
 class ManhattanHeuristicAlgorithm implements DistanceAlgorithm{
 	
-	public function calculate(Tile $from, Tile $to) : int{
+	public function calculate(Tile $from, Tile $to) : float{
 		return abs($from->x - $to->x) + abs($from->y - $to->y);
 	}
 }
